# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TinyURL Shop by Housni BENABID",
        "description": "Important! For both index queries (Product, Category) the first request is performed from DB. The application generates Redis cache for that request so it will load faster.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories/{uuid}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Show product",
                "description": "Show product endpoint",
                "operationId": "category.show",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Category uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update product",
                "description": "Create product endpoint",
                "operationId": "category.destroy",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Category uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update product",
                "description": "Update product endpoint",
                "operationId": "category.update",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Category uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update product request body",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The product name",
                                        "type": "string",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List products",
                "description": "List products endpoint",
                "operationId": "products.index",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Field to sort by",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "top",
                        "in": "query",
                        "description": "Number of top results to return",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category filter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/products/": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create product",
                "description": "Create product endpoint",
                "operationId": "product.create",
                "requestBody": {
                    "description": "Create product request body",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price",
                                    "top",
                                    "categories"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The product name",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "description": {
                                        "description": "The description of the product",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "price": {
                                        "description": "The price of the product",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "top": {
                                        "description": "The price of the product",
                                        "type": "boolean",
                                        "default": "false"
                                    },
                                    "categories": {
                                        "description": "Array of categories UUID",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "default": "[]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/products/{uuid}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Show product",
                "description": "Show product endpoint",
                "operationId": "product.show",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Product uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Update product",
                "description": "Create product endpoint",
                "operationId": "product.destroy",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Product uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Products"
                ],
                "summary": "Update product",
                "description": "Update product endpoint",
                "operationId": "product.update",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Product uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update product request body",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price",
                                    "top",
                                    "categories"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The product name",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "description": {
                                        "description": "The description of the product",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "price": {
                                        "description": "The price of the product",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "top": {
                                        "description": "The price of the product",
                                        "type": "boolean",
                                        "default": "false"
                                    },
                                    "categories": {
                                        "description": "Array of categories UUID",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "default": "[]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/last-viewed-products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List last viwed products",
                "description": "List last viwed products endpoint",
                "operationId": "products.lastviewed",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Error processing the request",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    },
...
}